frontend_dialogpackage
1.封装登录注册dialog
2.修改登录后用户情报显示及内容
